requirements.txt
h5py == 2.10.0
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.1.5
scikit-learn==0.24.0
tensorboard==1.14.0
tensorflow-gpu==1.14.0
tensorflow-estimator==1.14.0

